Move NestedSelect.jsx to the components/Filters folder
